refactor: extract playground CLI runner helpers
